+++ Mock-Bank-App/Withdraw.cs
@@ -11,7 +11,7 @@ namespace Mock_Bank_App
     internal class Withdraw
     {
 
-        private float userWithdrawAmount;
+        private decimal userWithdrawAmount;
         private bool userEnteredValidWithdrawlAmount = false;
         private TransactionHistory _history;
         
@@ -20,11 +20,11 @@ namespace Mock_Bank_App
             _history = history;
         }
         //Allows the user to withdraw funds from their account. This calls upon a CheckAmountAgainstBalance method to make sure the transaction is valid
-        public float Transaction(float currentAmount)
+        public decimal Transaction(decimal currentAmount)
         {
             ColoredConsole.WriteLineColored("Current balance: £" + currentAmount, ConsoleColor.Yellow);
             ColoredConsole.WriteLineColored("How much would you like to withdraw?", ConsoleColor.Blue);
-            userWithdrawAmount = Convert.ToSingle(Console.ReadLine());
+            userWithdrawAmount = Convert.ToDecimal(Console.ReadLine());
             while (userEnteredValidWithdrawlAmount == false)
             {
                 currentAmount = CheckAmountAgainstBalance(currentAmount);
@@ -36,9 +36,9 @@ namespace Mock_Bank_App
         }
 
         //This checks the amount being withdrawn and ensures the user does not drop below 0. If the amount withdrawn - the current balance = 0 then the user is informed they cannot do that
-        private float CheckAmountAgainstBalance(float currentAmount)
+        private decimal CheckAmountAgainstBalance(decimal currentAmount)
         {
-            if (currentAmount - userWithdrawAmount < 0) { ColoredConsole.WriteLineColored("You cannot withdraw that much. Please enter a different number.", ConsoleColor.Red); userWithdrawAmount = Convert.ToSingle(Console.ReadLine()); }
+            if (currentAmount - userWithdrawAmount < 0) { ColoredConsole.WriteLineColored("You cannot withdraw that much. Please enter a different number.", ConsoleColor.Red); userWithdrawAmount = Convert.ToDecimal(Console.ReadLine()); }
             else { currentAmount -= userWithdrawAmount; ColoredConsole.WriteLineColored("New balance: £" + currentAmount, ConsoleColor.Yellow); userEnteredValidWithdrawlAmount = true; _history.AddToTransactionList(userWithdrawAmount, currentAmount, "withdraw"); }
 
             return currentAmount;
